Minor Fixes
-- Refactors the choose target code slightly to make it a bit more readable
-- Adds proper battle constants for the defense calculation
-- Modifies the TODO list for up-to-date info

+++ src/constants/BattleConstants.java
@@ -10,7 +10,7 @@ public class BattleConstants {
     // when defense is 10% of health, defend effectiveness is halved
     public static final double DEFENSE_HALF_LIFE_HP_RATIO = 0.15;
     // minimum defense that occurs with defend action
-    public static final double MINIMUM_DEFENSE = 1;
+    public static final int MINIMUM_DEFENSE = 1;
 
     // Action Type Data
